Decompose process.rs into smaller modules for 400-line limit
- Create trait_impl.rs module with LanguageServer trait implementation
- Move impl LanguageServer for ProcessLanguageServer to trait_impl.rs
- Make helper methods pub(super) in process.rs for trait_impl access
- Add set_running_state() method to encapsulate state updates
- Add mod trait_impl declaration in mod.rs
- process.rs is now 465 lines (down from 615)
- All tests pass and formatting checks succeed
